Calculate a trimmed mean, removing n values from each end of the sequence

Calculate a trimmed mean, removing n values from each end of the sequence

mean_trim_by_n(x, n.trim)

Arguments

x

A numeric vector

n.trim

A number

Value

The vector mean with 'n.trim' values removed from each end before calculating